moved neuron logging to log_utils
